requirements.txt
# 核心依赖
numpy>=1.20.0
h5py>=3.1.0
pyyaml>=6.0
matplotlib>=3.5.0
pandas>=1.3.0
tqdm>=4.62.0

# 数据库客户端
pymilvus>=2.2.0
psycopg2-binary>=2.9.0
pg-vector>=0.1.0

# 可选依赖
scikit-learn>=1.0.0  # 用于数据处理和分析
plotly>=5.5.0  # 用于高级可视化

ansicolors==1.1.8
psutil==5.9.4
jinja2==3.1.2
pytest==7.2.2
datasets==2.12.0
requests==2.31.0




